% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildAncDesCoordDF}
\alias{buildAncDesCoordDF}
\title{Returns the coordinate positions of all ancestors and descendants of a variety}
\usage{
buildAncDesCoordDF(df)
}
\arguments{
  \item{df}{data frame of the ancestors and descendants of
  a variety (from function buildAncDesTotalDF)}
}
\description{
Calculates coordinates to plot each ancestors and
descendant of a variety on a tree. The x and y values
describe the coordinates of the label, while the xstart,
ystart, xend, and yend values describe the edges of the
label.
}
\seealso{
\code{\link{getAncestors}} for information on determining
ancestors

\code{\link{getDescendants}} for information on determining
descendants
}

