\name{buildDegMatrix}
\alias{buildDegMatrix}
\title{Returns the image object to show the heat map of degrees between the inputted set of vertices}
\usage{
buildDegMatrix(varieties)
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}
}
\description{
Returns the image object to show the heat map of degrees between the inputted set of vertices
}

