\name{buildYearMatrix}
\alias{buildYearMatrix}
\title{Returns the image object to show the heat map of years between the inputted set of vertices}
\usage{
buildYearMatrix(varieties)
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}
}
\description{
Returns the image object to show the heat map of years between the inputted set of vertices
}

