% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAncestors}
\alias{getAncestors}
\title{Returns a list of the ancestors of a particular variety (if they exist)}
\usage{
getAncestors(v1, tree, gen = 3)
}
\arguments{
  \item{v1}{the variety}

  \item{tree}{the tree}

  \item{gen}{the number of generations back to include as
  ancestors}
}
\description{
This function returns a list of the ancestors of the
inputted variety within and including a given number of
generations
}
\examples{
data(sbTree)
getParent("Essex", sbTree)
getAncestors("Essex", sbTree, 1)
getAncestors("Essex", sbTree, 5)
}

