% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBasicStatistics}
\alias{getBasicStatistics}
\title{Determine basic statistics of the graph object}
\usage{
getBasicStatistics(ig)
}
\arguments{
  \item{ig}{the igraph representation of the tree}
}
\description{
Returns basic statistics of the graph object (number of
nodes, number of edges, whether or not the whole graph is
connected, number of components, average path length, graph
diameter, etc.)
}
\examples{
getBasicStatistics(ig)
}

