% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getChild}
\alias{getChild}
\title{Returns the children of a particular variety (if they exist)}
\usage{
getChild(v1, tree)
}
\arguments{
  \item{v1}{the first variety}

  \item{tree}{the tree}
}
\description{
This function returns zero or more values that indicate the
children of the inputted variety.
}
\examples{
data(sbTree)
getChild("Tokyo", sbTree)
getChild("Essex", sbTree)
}

