% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDescendants}
\alias{getDescendants}
\title{Returns a list of the descendants of a particular variety (if they exist)}
\usage{
getDescendants(v1, tree, gen = 3)
}
\arguments{
  \item{v1}{the variety}

  \item{tree}{the tree}

  \item{gen}{the number of generations back to include as
  descendants}
}
\description{
This function returns a list of the descendants of the
inputted variety within and including a given number of
generations
}
\examples{
data(sbTree)
getChild("Essex",sbTree)
getDescendants("Essex", sbTree, 1)
getDescendants("Essex", sbTree, 3)
}

