% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEdges}
\alias{getEdges}
\title{Returns edges (vertex names and edge weights) for a tree}
\usage{
getEdges(ig)
}
\arguments{
  \item{ig}{iGraph object}
}
\description{
Returns a matrix, where each row contains information about
an edge (two vertex names and edge weight, if present) of
the tree.
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
getEdges(ig)
}

