% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNodes}
\alias{getNodes}
\title{Returns the nodes for a tree}
\usage{
getNodes(tree)
}
\arguments{
  \item{tree}{the tree object}
}
\description{
Returns a character list, where roww contains names of the
unique nodes in a tree.
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
getNodes(ig)
}

