% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getParent}
\alias{getParent}
\title{Returns the parents of a particular variety (if they exist)}
\usage{
getParent(v1, tree)
}
\arguments{
  \item{v1}{the first variety}

  \item{tree}{the tree}
}
\description{
This function returns up to two values that indicate the
parents of the inputted variety.
}
\examples{
data(sbTree)
getParent("Tokyo", sbTree)
getParent("Essex", sbTree)
}

