% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getYear}
\alias{getYear}
\title{Determine the year of a variety}
\usage{
getYear(v1, tree)
}
\arguments{
  \item{v1}{the variety}

  \item{tree}{tree}
}
\description{
Returns the documented year of the inputted variety
}
\examples{
data(sbTree)
getYear("Essex",sbTree)
getYear("Tokyo",sbTree)
}

