% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPathOnTree}
\alias{plotPathOnTree}
\title{Build the image object of the whole tree with the path superimposed onto it}
\usage{
plotPathOnTree(path, ig, binVector = sample(1:12, 12))
}
\arguments{
  \item{path}{path as returned from getPath() or a vector
  of two variety names which exist in ig}

  \item{ig}{igraph representation of the tree}

  \item{binVector}{vector of numbers between 1 and
  length(binVector), each repeated exactly once}
}
\description{
This function requires a path and the ig object, and plots
the entire tree with the path highlighted. The image will
correctly position the node labels with x-axis representing
the node year, and y-axis representing the node path index.
Light grey edges between two nodes represent parent-child
relationships between those nodes. To enhance the visual
understanding of how the path-of-interest fits into the
entire graph structure, the nodes within the path are
labelled in boldface, and connected with light-green
boldfaced edges.
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
path = getPath("Brim","Bedford",ig,sbTree)
plotTotalImage <- plotPathOnTree(path=path, ig=ig)
plotTotalImage
}
\seealso{
\url{http://www.r-project.org} for iGraph information

\code{\link{getPath}} for information on input path
building
}

