% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treeToIG}
\alias{treeToIG}
\title{Process the tree graph}
\usage{
treeToIG(tree, vertexinfo = NULL, edgeweights = 1, isDirected = FALSE)
}
\arguments{
  \item{tree}{A data frame representing tree information,
  containing one row for each EDGE, with at least two
  columns named parent and child, respectively,
  representing vertices connected by an edge. Terminal
  nodes should have "NA" for parent in order to preserve
  vertex information.}

  \item{vertexinfo}{(default NULL) either names of columns
  in the tree which should be added to the database as
  vertex information or a data frame with information for
  all vertices such that the first column contains vertex
  names.}

  \item{edgeweights}{(default 1) name of a column which
  contains edge weights}

  \item{isDirected}{(default FALSE) should the graph be a
  directed graph?}
}
\description{
Processes the tree into an igraph object with appropriate
vertex information, graph type, and edge weights.
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

