% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildAncDesTotalDF}
\alias{buildAncDesTotalDF}
\title{Returns the data frame of labels and plot coordinates of all ancestors and descendants of a variety}
\usage{
buildAncDesTotalDF(v1, tree, mAnc = 3, mDes = 3)
}
\arguments{
  \item{v1}{variety of interest}

  \item{tree}{the tree}

  \item{mAnc}{maximum number of ancestors of v1 to be
  shown}

  \item{mDes}{maximum number of descendants of v1 to be
  shown}
}
\description{
Returns the data frame that includes labels and plot
coordinates of all ancestors and descendants of a variety.
Users can specify the maximum number of ancestors and
descendants to display
}
\examples{
data(sbTree)
tree=sbTree
v1="Essex"
buildAncDesTotalDF("Essex", tree)
}
\seealso{
\code{\link{getAncestors}} for information on determining
ancestors

\code{\link{getDescendants}} for information on determining
descendants
}

