% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildEdgeTotalDF}
\alias{buildEdgeTotalDF}
\title{Build the edges in the  total tree graph}
\usage{
buildEdgeTotalDF(ig, binVector = 1:12)
}
\arguments{
  \item{ig}{igraph object from processTreeGraph}

  \item{binVector}{vector of numbers between 1 and
  length(binVector), each repeated exactly once}
}
\description{
This function takes the ig object and creates a data frame
object of the edges between all parent-child relationships
in the graph
}

