% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildMinusPathDF}
\alias{buildMinusPathDF}
\title{Process the tree graph}
\usage{
buildMinusPathDF(path, ig, binVector = 1:12)
}
\arguments{
  \item{path}{path as returned from getPath() or a vector
  of two variety names which exist in the ig object}

  \item{ig}{the igraph representation of the tree}

  \item{binVector}{vector of numbers between 1 and
  length(binVector), each repeated exactly once}
}
\description{
This function takes the spreadTotalDF object (from the
buildSpreadTotalDF function) and the path object as inputs.
From these objects, it creates a data frame object of the
label, x, and y values of all nodes in the tree. However,
the data frame object does not include the labels of the
path varieties, as they will be treated differently.
}
\seealso{
\url{http://www.r-project.org} for iGraph information

\code{\link{getPath}} for information on input path
building
}

