% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateGenPlot}
\alias{generateGenPlot}
\title{Returns the image object to show the ancestors and descendants of a variety}
\usage{
generateGenPlot(gDF)
}
\arguments{
  \item{genDF}{data frame created from the buildGenDF
  function}
}
\description{
Returns the image object to show the ancestors (to the
left) and descendants (to the right) of a variety, with the
variety highlighted in orange
}
\seealso{
\code{\link{getancestors}} for information on determining
ancestors

\code{\link{getdescendants}} for information on determining
descendants
}

