% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAncestors}
\alias{getAncestors}
\title{Returns the ancestors of a particular variety (if they exist)}
\usage{
getAncestors(v1, gen = 0)
}
\arguments{
  \item{v1}{the first variety}

  \item{gen}{generation}
}
\description{
This function returns a nested list of the ancestors of the
inputted variety.
}
\seealso{
\code{\link{getParent}} for information on determining
parents
}

