% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getancestors}
\alias{getancestors}
\title{Returns the ancestors of a particular variety (if they exist)}
\usage{
getancestors(v1, gen = 0)
}
\arguments{
  \item{v1}{the first variety}

  \item{gen}{generation}
}
\description{
This function returns a nested list of the ancestors of the
inputted variety.
}
\seealso{
\code{\link{getparent}} for information on determining
parents
}

