\name{getBasicStatistics}
\alias{getBasicStatistics}
\title{Determine basic statistics of the graph object}
\usage{
getBasicStatistics(mygraph)
}
\arguments{
  \item{mygraph}{graph representation of the tree}
}
\description{
Returns basic statistics of the graph object (number of nodes, number of edges, whether or not the
whole graph is connected, number of components, average path length, graph diameter, etc.)
}
\examples{
getBasicStatistics(mygraph)
}

