% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDegree}
\alias{getDegree}
\title{Determine the degree between two varieties}
\usage{
getDegree(v1, v2, ig)
}
\arguments{
  \item{v1}{the first variety}

  \item{v2}{the second variety}

  \item{ig}{the igraph object}
}
\description{
Returns the degree (distance between unweighted edges)
between two varieties, where an edge represents a
parent-child relationship
}
\examples{
getDegree("Brim","Bedford",ig)
}

