% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdescendants}
\alias{getdescendants}
\title{Returns the descendants of a particular variety (if they exist)}
\usage{
getdescendants(v1, gen = 0)
}
\arguments{
  \item{v1}{the first variety}

  \item{gen}{generation the first variety}
}
\description{
This function returns a nested list of the descendants of
the inputted variety.
}
\seealso{
\code{\link{getchild}} for information on determining
parents
}

