\name{getparent}
\alias{getparent}
\title{Returns the parents of a particular variety (if they exist)}
\usage{
getparent(v1, tree)
}
\arguments{
  \item{v1}{the first variety}
}
\description{
This function returns up to two values that indicate the parents of the inputted variety.
}
\examples{
getparent("Tokyo")
getParent("Essex")
}

