% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPath}
\alias{getPath}
\title{Determine the degree between two varieties}
\usage{
getPath(v1, v2, ig, tree, silent = FALSE, isDirected = FALSE)
}
\arguments{
  \item{v1}{the first variety}

  \item{v2}{the second variety}

  \item{ig}{the igraph representation of the tree}

  \item{tree}{the tree}

  \item{silent}{Print output? Defaults to FALSE.}

  \item{isDirected}{boolean whether or not the graph is
  directed, defaults to FALSE}
}
\description{
Determines the shortest path between the two inputted
vertices, and takes into account whether or not the graph
is directed. If there is a path, the list of vertices of
the path will be returned. If there is not a path, a list
of character(0) will be returned. Note: For a directed
graph, the direction matters. However, this function will
check both directions and return the path if it exists.
}
\examples{
getPath("Brim","Bedford",ig)
getPath("Tokyo","Volstate",ig)
}

