% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isChild}
\alias{isChild}
\title{Determine if a variety is a child of another}
\usage{
isChild(child, parent, tree)
}
\arguments{
  \item{child}{possible child variety}

  \item{parent}{possible parent variety}

  \item{tree}{tree}
}
\description{
Returns a boolean variable for whether the first variety is
a child of the second variety
}
\examples{
data(sbTree)
isChild("Essex","Young",sbTree)
isChild("Young","Essex",sbTree)
}

