% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAncDes}
\alias{plotAncDes}
\title{Returns the image object to show the ancestors and descendants of a variety}
\usage{
plotAncDes(gDF)
}
\arguments{
  \item{genDF}{data frame created from the
  buildAncDesTotalDF function}
}
\description{
Returns the image object to show the ancestors (to the
left) and descendants (to the right) of a variety, with the
variety highlighted in orange
}
\examples{
data(sbTree)
adDF = buildAncDesTotalDF("Essex", sbTree)
plotAncDes(adDF)

data(sbTree)
adDF = buildAncDesTotalDF("Tokyo", sbTree)
plotAncDes(adDF)
}
\seealso{
\code{\link{getAncestors}} for information on determining
ancestors

\code{\link{getDescendants}} for information on determining
descendants
}

