% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAncDes}
\alias{plotAncDes}
\title{Returns the image object to show the ancestors and descendants of a variety}
\usage{
plotAncDes(v1, tree, mAnc = 3, mDes = 3, vColor = "#D35C79")
}
\arguments{
  \item{v1}{variety of interest}

  \item{tree}{the tree}

  \item{mAnc}{maximum number of ancestors of v1 to be
  shown}

  \item{mDes}{maximum number of descendants of v1 to be
  shown}

  \item{vColor}{the color of the text of the main variety}
}
\description{
Returns the image object to show the ancestors and
descendants of a variety, with the variety highlighted, if
desired
}
\examples{
data(sbTree)
plotAncDes("Essex", sbTree, 2, 3, "blue") + ggplot2::labs(x="Generation index",y="")
plotAncDes("Tokyo", sbTree, vColor="red")
}

