% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAncDes}
\alias{plotAncDes}
\title{Returns the image object to show the ancestors and descendants of a variety}
\usage{
plotAncDes(v1, tree, mAnc = 3, mDes = 3, colour = color)
}
\arguments{
  \item{v1}{variety of interest}

  \item{tree}{the tree}

  \item{mAnc}{maximum number of ancestors of v1 to be
  shown}

  \item{mDes}{maximum number of descendants of v1 to be
  shown}
}
\description{
Returns the image object to show the ancestors (to the
left) and descendants (to the right) of a variety, with the
variety highlighted in orange
}
\examples{
data(sbTree)
plotAncDes("Essex", sbTree, 5, 3)

data(sbTree)
plotAncDes("Tokyo", sbTree)
}

