% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDegMatrix}
\alias{plotDegMatrix}
\title{Returns the image object to show the heat map of degrees between the inputted set of vertices}
\usage{
plotDegMatrix(varieties, ig, tree, xLab = "Variety", yLab = "Variety",
  legendLab = "Degree")
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}

  \item{ig}{igraph}

  \item{tree}{tree}

  \item{xLab}{string label on the x axis (default is
  "Variety")}

  \item{yLab}{string label on the y axis (default is
  "Variety")}

  \item{legendLab}{string label on the legend (default is
  "Degree")}
}
\description{
Returns the image object to show the heat map of degrees
between the inputted set of vertices
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
varieties=c("Beeson", "Calland", "Dillon", "Hood", "Narow", "Pella", "Tokyo", "Young", "Zane")
plotDegMatrix(varieties,ig,sbTree,"Soybean name", "Soybean name", "Deg")

plotDegMatrix(varieties,ig, sbTree) + ggplot2::scale_fill_continuous(low="yellow", high="purple")
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

