% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDegMatrix}
\alias{plotDegMatrix}
\title{Returns the image object to show the heat map of degrees between the inputted set of vertices}
\usage{
plotDegMatrix(varieties, ig, tree)
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}

  \item{ig}{igraph}

  \item{tree}{tree}
}
\description{
Returns the image object to show the heat map of degrees
between the inputted set of vertices
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
varieties=c("Beeson", "Calland", "Dillon", "Hood", "Narow", "Pella", "Tokyo", "Young", "Zane")
plotDegMatrix(varieties,ig,sbTree)
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

