% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotYearMatrix}
\alias{plotYearMatrix}
\title{Returns the image object to show the heat map of years between the inputted set of vertices}
\usage{
plotYearMatrix(varieties, tree)
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}

  \item{tree}{tree}
}
\description{
Returns the image object to show the heat map of years
between the inputted set of vertices
}

