% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotYearMatrix}
\alias{plotYearMatrix}
\title{Returns the image object to show the heat map of years between the inputted set of vertices}
\usage{
plotYearMatrix(varieties, tree, xLab = "Variety", yLab = "Variety",
  legendLab = "Difference in years")
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}

  \item{tree}{tree}

  \item{xLab}{string label on the x axis (default is
  "Variety")}

  \item{yLab}{string label on the y axis (default is
  "Variety")}

  \item{legendLab}{string label on the legend (default is
  "Degree")}
}
\description{
Returns the image object to show the heat map of years
between the inputted set of vertices
}
\examples{
data(sbTree)
varieties=c("Bedford", "Calland", "Narow", "Pella", "Tokyo", "Young", "Zane")
p = plotYearMatrix(varieties,sbTree,"Variety", "Variety", "Degree")
p + ggplot2::scale_fill_continuous(low="white", high="darkgreen")
}

